backend/requirements.txt
# Hotel Munich Backend - Requirements
# =====================================
# FastAPI + Database + Security stack

# Core Database
sqlalchemy
pydantic
pydantic-settings>=2.0.0

# FastAPI Stack
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Security (JWT + Password Hashing + Rate Limiting)
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
slowapi>=0.1.9

# Utilities
python-dotenv>=1.2.1
requests>=2.32.5

# Logging
colorama>=0.4.6

# iCal Sync (Booking.com / Airbnb)
icalendar>=6.0.0

# AI (Google Gemini + Reliability)
google-genai>=1.0.0
tenacity>=8.2.0

# Testing (QA Skill)
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
